Ticket: #6074 - angle measurements and calibration are wrong on images where aspect ratio is not an integer '''Schema/Configuration Changes: N/A'''  * Added unit tests.  * Fixed calibration tool.  * Fixed angle measurement tool.

+++ ImageViewer/Tools/Measurement/ProtractorAnalyzers.cs
@@ -72,7 +72,7 @@ namespace ClearCanvas.ImageViewer.Tools.Measurement
 			}
 			else
 			{
-				aspectRatio = (float)(roiInfo.PixelAspectRatio.Row/roiInfo.PixelAspectRatio.Column);
+				aspectRatio = (float)roiInfo.PixelAspectRatio.Row / (float)roiInfo.PixelAspectRatio.Column;
 			}
 
 			List<PointF> normalized = new List<PointF>();
